refactor: nextjs의 router groups를 활용한 폴더 구조 변경
