+++ app/(user)/apply/page.tsx
