+++ app/(user)/certificate/page.tsx
@@ -1,0 +1,5 @@
+import Certificate from '@/components/Certificate';
+
+const CertificatePage = () => <Certificate />;
+
+export default CertificatePage;
