--- app/certificate/page.tsx
@@ -1,5 +1,0 @@
-import Certificate from '@/components/Certificate';
-
-const ApplyPage = () => <Certificate />;
-
-export default ApplyPage;
