+++ hooks/index.tsx
@@ -6,3 +6,4 @@ export * from './useLogin';
 export * from './useInputRef';
 export * from './useDeviceChecker';
 export * from './useBeforeUnload';
+export * from './useMousePosition';
